+++ src/clipboard.rs
@@ -1,4 +1,4 @@
-use clippers::Clipboard;
+use arboard::Clipboard;
 use std::fs::File;
 use std::io::Read;
 use std::path::Path;
@@ -10,9 +10,9 @@ pub fn copy_to_clipboard(output_path: &Path) -> AppResult<()> {
     let mut output_contents = String::new();
     File::open(output_path)?.read_to_string(&mut output_contents)?;
 
-    let mut clipboard = Clipboard::get();
+    let mut clipboard = Clipboard::new().map_err(|e| AppError::Clipboard(e.to_string()))?;
     clipboard
-        .write_text(&output_contents)
+        .set_text(output_contents)
         .map_err(|e| AppError::Clipboard(e.to_string()))?;
 
     Ok(())