+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 ## 1.7.2 - 2026-02-09
 
 Changed
-- Replaced `clipboard` with `clippers` to avoid the `xcb` dependency.
+- Replaced `clipboard` with `arboard` to avoid the `xcb` dependency and CMake build step.
 
 ## 1.7.1 - 2026-02-09